requirements.txt
celery==5.2.3
redis==3.5.3
flask==2.1.1
scikit-image
Pillow==9.1.0